requirements.txt
torch
torchvision
streamlit
opencv-python
Pillow
matplotlib
ultralytics
